requirements.txt
Django==4.2.23
djangorestframework==3.15.2
django-filter==24.3
psycopg2-binary==2.9.10
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.0.1 